requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==21.4.0
blinker==1.9.0
boto3==1.21.27
botocore==1.24.27
cachetools==5.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.7
cloudpickle==2.1.0
cycler==0.11.0
dice-ml==0.7.2
distro==1.9.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exa_py==1.4.0
exceptiongroup==1.2.2
filelock==3.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.31.2
fsspec==2024.10.0
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.6.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub>=0.17.0
idna==3.3
imageio==2.16.1
inflect==5.4.0
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.4.0
lark==1.1.2
lime==0.2.0.1
llvmlite==0.39.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
nltk==3.7
numba==0.56.0
numpy==1.21.6
openai==1.52.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandas-stubs==1.2.0.53
pathtools==0.1.2
patsy==0.5.2
Pillow==9.0.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.0.0
promise==2.3
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.1
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rsa==4.9
s3transfer==0.5.2
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers>=2.2.2
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.2
shap==0.40.0
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
statsmodels==0.13.2
tenacity==9.0.0
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.20.1
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.63.1
transformers>=4.30.0
twilio==7.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.32.1
wandb==0.18.5
Werkzeug==3.1.3
word2number==1.1
wordninja==2.0.0
psycopg2-binary==2.9.9